Rename and move files in manifest command to match the rest of the tool structure.
